MoleculeSpectrum.jl API Documentation

Contents

MoleculeSpectrum.jl API Documentation

Contents

Molecular Parameters

Molecular Parameters

Polarizability Parameters

Zeeman Parameters

Nuclear Parameters

Vectors and Tensors

Spherical vectors

Unit vectors

Tensors

States

Types

Functions

Fields

Hamiltonians

Spectra and Analysis

Index

Molecular Parameters

Molecular Parameters

MoleculeSpectrum.MolecularParameters — Type

struct MolecularParameters
    "Permanent dipole moment (Debye)"
    dₚ::Float64
    "Rotational constant (MHz)"
    Bᵣ::Float64
    "Nuclear angular momenta"
    I::SVector{2,HalfInt}
    "Zeeman parameters"
    zeeman::ZeemanParameters
    "Nuclear Parameters"
    nuclear::NuclearParameters
    "Molecular polarizability at the trapping wavelength"
    α::Polarizability
end

Contains the coupling constants for the molecular Hamiltonian and the nuclear angular momenta (needed to construct the basis states).

MoleculeSpectrum.KRb_Parameters_Neyenhuis — Constant

KRb_Parameters_Neyenhuis = MolecularParameters(
    0.574,
    1113.9514,
    [HalfInt(4), HalfInt(3 / 2)],
    KRb_Zeeman,
    KRb_Nuclear_Neyenhuis,
    KRb_Polarizability,
)

MolecularParameters with experimental values from Neyenhuis et al., PRL 109, 230403 (2012)

MoleculeSpectrum.KRb_Parameters_Ospelkaus — Constant

KRb_Parameters_Ospelkaus = MolecularParameters(
    0.574,
    1113.950,
    [HalfInt(4), HalfInt(3 / 2)],
    KRb_Zeeman,
    KRb_Nuclear_Ospelkaus,
    KRb_Polarizability,
)

MolecularParameters with experimental values from Ospelkaus et al., PRL 104, 030402 (2010)

MoleculeSpectrum.DEFAULT_MOLECULAR_PARAMETERS — Constant

DEFAULT_MOLECULAR_PARAMETERS

Default molecular parameters; alias for KRb_Parameters_Neyenhuis

MoleculeSpectrum.TOY_MOLECULE_PARAMETERS — Constant

TOY_MOLECULE_PARAMETERS

Toy model values with dipole = 1 D, no hyperfine structure, etc. Intended for testing.

Note that the formulas break down for I = 0, which is why we use I = 1 here.

Polarizability Parameters

MoleculeSpectrum.Polarizability — Type

struct Polarizability
    "Parallel ac polarizability (MHz / (W / cm^2))"
    α_par::Float64
    "Perpendicular ac polarizability (MHz / (W / cm^2))"
    α_perp::Float64
end

Contains the parallel and perpendicular ac polarizabilities at a particular optical wavelength λ.

So far, it is assumed that λ is far-detuned from any electronic transitions, such that the polarizability does not depend on rotational state $N$.

MoleculeSpectrum.KRb_Polarizability — Constant

KRb_Polarizability = Polarizability(10.0e-5, 3.3e-5)

Polarizability with experimental values from Neyenhuis et al., PRL 109, 230403 (2012)

Zeeman Parameters

MoleculeSpectrum.ZeemanParameters — Type

struct ZeemanParameters
    "Rotational g factor"
    gᵣ::Float64
    "Nuclear g factor"
    gᵢ::SVector{2,Float64}
    "Nuclear shielding factor"
    σᵢ::SVector{2,Float64}
end

Contains the g-factors and nuclear shielding factors for computing Zeeman shifts.

MoleculeSpectrum.KRb_Zeeman — Constant

KRb_Zeeman = ZeemanParameters(0.014, [-0.324, 1.834], [1321e-6, 3469e-6])

ZeemanParameters with theoretical values from Aldegunde et al., PRA 78, 033434 (2008)

Nuclear Parameters

MoleculeSpectrum.NuclearParameters — Type

struct NuclearParameters
    "Nuclear electric quadrupole (MHz)"
    eqQᵢ::SVector{2,Float64}
    "Nuclear spin-rotation interaction (MHz)"
    cᵢ::SVector{2,Float64}
    "Nuclear spin-spin scalar interaction (MHz)"
    c₄::Float64
end

Contains the nuclear electric quadrupole moments, nuclear spin-rotation couplings, and nuclear spin-spin coupling for calculating the hyperfine Hamiltonian.

MoleculeSpectrum.KRb_Nuclear_Neyenhuis — Constant

KRb_Nuclear_Neyenhuis = NuclearParameters([0.45, -1.308], [-24.1e-6, 420.1e-6], -2030.4e-6)

NuclearParameters with experimental values from Neyenhuis et al., PRL 109, 230403 (2012)

MoleculeSpectrum.KRb_Nuclear_Ospelkaus — Constant

KRb_Nuclear_Ospelkaus = NuclearParameters([0.45, -1.41], [-24.1e-6, 420.1e-6], -2030.4e-6)

NuclearParameters with experimental values from Ospelkaus et al., PRL 104, 030402 (2010)

Vectors and Tensors

Spherical vectors

MoleculeSpectrum.SphericalVector — Type

SphericalVector(magnitude, θ, φ)

Construct a vector with magnitude, polar angle θ, and azimuthal angle φ.

Represents an external field vector in spherical coordinates used to construct ExternalFields. Currently, SphericalVectors may be negated but no other mathematical operations are implemented.

Vectors along $x$, $y$, or $z$ can be quickly constructed using VectorX, VectorY, and VectorZ, respectively.

See also SphericalUnitVector, ExternalFields.

MoleculeSpectrum.VectorX — Function

VectorX(magnitude)

Construct a SphericalVector with magnitude along x.

Examples

julia> VectorX(5.0)
SphericalVector(5.0, 1.5707963267948966, 0.0)

MoleculeSpectrum.VectorY — Function

VectorY(magnitude)

Construct a SphericalVector with magnitude along y.

Examples

julia> VectorY(2.0)
SphericalVector(2.0, 1.5707963267948966, 1.5707963267948966)

MoleculeSpectrum.VectorZ — Function

VectorZ(magnitude)

Construct a SphericalVector with magnitude along z.

Examples

julia> VectorZ(10.0)
SphericalVector(10.0, 0.0, 0.0)

Unit vectors

MoleculeSpectrum.SphericalUnitVector — Type

SphericalUnitVector(magnitude, θ, φ)
SphericalUnitVector(v::SphericalVector)

Construct a unit vector with polar angle θ, and azimuthal angle φ.

Represents the direction of an external field vector in spherical coordinates. Currently, SphericalUnitVectors may be negated but no other mathematical operations are implemented.

Vectors along $x$, $y$, or $z$ can be quickly constructed using UnitVectorX, UnitVectorY, and UnitVectorZ, respectively.

See also SphericalVector, find_transition_strengths, T⁽¹⁾, T⁽²⁾.

MoleculeSpectrum.UnitVectorX — Function

UnitVectorX()

Construct a SphericalUnitVector along x.

Examples

julia> UnitVectorX()
SphericalUnitVector(1.5707963267948966, 0.0)

MoleculeSpectrum.UnitVectorY — Function

UnitVectorY()

Construct a SphericalUnitVector along y.

Examples

julia> UnitVectorY()
SphericalUnitVector(1.5707963267948966, 1.5707963267948966)

MoleculeSpectrum.UnitVectorZ — Function

UnitVectorZ()

Construct a SphericalUnitVector along z.

Examples

julia> UnitVectorZ()
SphericalUnitVector(0.0, 0.0)

Tensors

MoleculeSpectrum.T⁽¹⁾ — Function

T⁽¹⁾(v)

Construct the components of the rank 1 spherical tensor $T⁽¹⁾(v)$.

Examples

julia> T⁽¹⁾(UnitVectorX())
3-element StaticArrays.SVector{3, ComplexF64} with indices SOneTo(3):
    0.7071067811865475 - 0.0im
 6.123233995736766e-17 + 0.0im
   -0.7071067811865475 - 0.0im

julia> T⁽¹⁾(UnitVectorY())
3-element StaticArrays.SVector{3, ComplexF64} with indices SOneTo(3):
  4.329780281177466e-17 - 0.7071067811865475im
  6.123233995736766e-17 + 0.0im
 -4.329780281177466e-17 - 0.7071067811865475im

julia> T⁽¹⁾(UnitVectorZ())
3-element StaticArrays.SVector{3, ComplexF64} with indices SOneTo(3):
  0.0 - 0.0im
  1.0 + 0.0im
 -0.0 - 0.0im

MoleculeSpectrum.T⁽²⁾ — Function

T⁽²⁾(v)

Construct the components of the rank 2 spherical tensor $T⁽²⁾(v, v)$.

Examples

julia> T⁽²⁾(UnitVectorX())
5-element StaticArrays.SVector{5, ComplexF64} with indices SOneTo(5):
                    0.5 - 0.0im
  6.123233995736766e-17 - 0.0im
    -0.4082482904638631 + 0.0im
 -6.123233995736766e-17 - 0.0im
                    0.5 + 0.0im

julia> T⁽²⁾(UnitVectorY())
5-element StaticArrays.SVector{5, ComplexF64} with indices SOneTo(5):
                   -0.5 - 6.123233995736766e-17im
  3.749399456654644e-33 - 6.123233995736766e-17im
    -0.4082482904638631 + 0.0im
 -3.749399456654644e-33 - 6.123233995736766e-17im
                   -0.5 + 6.123233995736766e-17im

julia> T⁽²⁾(UnitVectorZ())
5-element StaticArrays.SVector{5, ComplexF64} with indices SOneTo(5):
                0.0 - 0.0im
                0.0 - 0.0im
 0.8164965809277261 + 0.0im
               -0.0 - 0.0im
                0.0 + 0.0im

MoleculeSpectrum.get_tensor_component — Function

get_tensor_component(p::Int, tensor)

#TODO: Document what get_tensor_component does.

MoleculeSpectrum.tensor_dot — Function

tensor_dot(a, b)

Contract two spherical tensors a and b.

States

Types

MoleculeSpectrum.State — Type

struct State
    N::Int
    mₙ::Int
    I::SVector{2,HalfInt}
    mᵢ::SVector{2,HalfInt}
end

Represents a molecular state in the uncoupled basis.

MoleculeSpectrum.KRbState — Function

KRbState(N, mₙ, mK, mRb)

Creates a basis state $|N, m_n, m_{\text{K}}, m_{\text{Rb}}⟩$ for ${}^{40}\text{K}^{87}\text{Rb}$.

This is a wrapper around State to avoid having to specify the nuclear spins $I_k$ each time.

See also State.

Functions

MoleculeSpectrum.index_to_state — Function

index_to_state(i, I₁, I₂)

Returns the State corresponding to the ith member of the basis.

The uncoupled basis $|N, mₙ, I₁, mᵢ₁, I₂, mᵢ₂⟩$ is ordered with the quantum numbers on the left changing the slowest.

See also State, state_to_index.

Examples

julia> s = index_to_state(1, 4, 3/2)
State(0, 0, HalfIntegers.Half{Int64}[4, 3/2], HalfIntegers.Half{Int64}[-4, -3/2])

julia> s = index_to_state(37, 4, 3/2)
State(1, -1, HalfIntegers.Half{Int64}[4, 3/2], HalfIntegers.Half{Int64}[-4, -3/2])

MoleculeSpectrum.state_to_index — Function

state_to_index(s::State)

Returns index of state s in the basis.

The uncoupled basis $|N, mₙ, I₁, mᵢ₁, I₂, mᵢ₂⟩$ is ordered with the quantum numbers on the left changing the slowest.

See also State, index_to_state.

Examples

julia> state_to_index(KRbState(1, 1, -4, 1/2))
111

julia> state_to_index(index_to_state(42, 4, 3/2))
42

MoleculeSpectrum.order_by_overlap_with — Function

order_by_overlap_with(spectrum, target)

Orders the list of eigenstates from spectrum by their wavefunction overlap with target.

See also calculate_spectrum, max_overlap_with.

MoleculeSpectrum.max_overlap_with — Function

max_overlap_with(spectrum, target)

Find the eigenstate from spectrum with the most wavefunction overlap with target.

Returns a tuple (overlap, index), where overlap is the wavefunction overlap with target and index is the position of the eigenstates in spectrum.

See also calculate_spectrum, order_by_overlap_with.

MoleculeSpectrum.find_closest_basis_state — Function

find_closest_basis_state(spectrum, index)

Find the basis state with the most wavefunction overlap with the eigenstate from spectrum at index.

This method can be thought of as an inverse to max_overlap_with.

See also calculate_spectrum, max_overlap_with.

MoleculeSpectrum.decompose_to_basis_states — Function

decompose_to_basis_states(spectrum::Spectrum, index::Int)
decompose_to_basis_states(spectrum::Spectrum, state::Vector{ComplexF64})
decompose_to_basis_states(state::Vector{ComplexF64}, I1, I2)

Decompose the given state into basis states in order of decreasing weight.

Outputs a Vector{Tuple{State, Float64}}, giving a list [(basis_state, weight)]. The first variant provides the decomposition of the member of spectrum.eigenstates at position index. The second and third variants provide the decomposition of the vector state. The third variant does not require spectrum and recomputes the basis set according to I1 and I2.

See also calculate_spectrum, max_overlap_with.

Fields

MoleculeSpectrum.ExternalFields — Type

ExternalFields(B::SphericalVector, E::SphericalVector, Optical::Vector{SphericalVector})
ExternalFields(B::Float64, E::Float64)

External magnetic, electric, optical fields to use in constructing the Hamiltonian.

If B and E are provided as Float64s, then the fields are assumed to be along z. The Optical argument can also be left as an empty vector [].

See also calculate_spectrum, hamiltonian, SphericalVector.

Examples

julia> ExternalFields(VectorZ(545.9), VectorX(1020.0), [])
ExternalFields(SphericalVector(545.9, 0.0, 0.0), SphericalVector(1020.0, 1.5707963267948966, 0.0), SphericalVector[])

julia> ExternalFields(545.9, 1020.0)
ExternalFields(SphericalVector(545.9, 0.0, 0.0), SphericalVector(1020.0, 0.0, 0.0), SphericalVector[])

julia> ExternalFields(VectorZ(545.9), VectorX(1020.0), [VectorY(2300.), SphericalVector(2300., π/2, π/4)])
ExternalFields(SphericalVector(545.9, 0.0, 0.0), SphericalVector(1020.0, 1.5707963267948966, 0.0), SphericalVector[SphericalVector(2300.0, 1.5707963267948966, 1.5707963267948966), SphericalVector(2300.0, 1.5707963267948966, 0.7853981633974483)])

MoleculeSpectrum.DEFAULT_FIELDS — Constant

DEFAULT_FIELDS = ExternalFields(545.9, 0.0)

Defines an ExternalFields with a magnetic field of 545.9 G along the z axis.

MoleculeSpectrum.TEST_FIELDS — Constant

TEST_FIELDS = ExternalFields(
    SphericalVector(545.9, π / 4, π / 4),
    SphericalVector(1000.0, 3π / 4, 7π / 4),
    [SphericalVector(2350.0, 0.0, 0.0)],
)

Defines an ExternalFields with a magnetic, electric, and optical fields for tests.

Hamiltonians

MoleculeSpectrum.hamiltonian — Function

hamiltonian(parts, external_fields)

Construct the full Hamiltonian including magnetic, electric, and optical fields.

Returns a dense Hermitian matrix.

The field-independent building blocks in parts can be reused over calls to hamiltonian to avoid recalculating the matrix elements each time.

See also make_hamiltonian_parts, make_krb_hamiltonian_parts, ExternalFields.

MoleculeSpectrum.HamiltonianParts — Type

struct HamiltonianParts
    basis::Vector{State}
    rotation::SparseHamiltonian
    dipole::SVector{3, SparseHamiltonian}
    dipole_relative::SVector{3, SparseHamiltonian} # used for transition strengths
    hyperfine::SparseHamiltonian
    zeeman::SVector{3, SparseHamiltonian}
    ac_scalar::SparseHamiltonian
    ac_tensor::SVector{5, SparseHamiltonian}
end

Contains all parts of the Hamiltonian except external fields.

Zeeman, dc Stark, and ac Stark terms are stored as vectors of matrices that can be contracted with the appropriate external field tensors. This allows the full Hamiltonian to be re-constructed at various field values and orientations without recalculating all of the matrix elements.

Should be created by make_hamiltonian_parts or make_krb_hamiltonian_parts. Used as an input to calculate_spectrum and hamiltonian.

See also make_hamiltonian_parts, make_krb_hamiltonian_parts.

MoleculeSpectrum.make_hamiltonian_parts — Function

make_hamiltonian_parts(molecular_parameters, N_max)

Construct all parts of the Hamiltonian that do not depend on external fields.

The size of the basis is determined by molecular_parameters, which contains the nuclear spin quantum numbers molecular_parameters.I, and the rotational states 0:N_max to include.

See also make_krb_hamiltonian_parts.

MoleculeSpectrum.make_krb_hamiltonian_parts — Function

make_krb_hamiltonian_parts(N_max)

Construct all parts of the ${}^{40}\text{K}^{87}\text{Rb}$ Hamiltonian that do not depend on external fields.

The rotational states 0:N_max are included. This is a shortcut method that replaces make_hamiltonian_parts for KRb.

See also make_hamiltonian_parts.

Spectra and Analysis

MoleculeSpectrum.Spectrum — Type

Spectrum(hamiltonian_parts, energies, eigenstates)

Represents a molecular spectrum, including eigenvalues and eigenstates, given HamiltonianParts and ExternalFields.

Created from calculate_spectrum. See decompose_to_basis_states and get_energyfor access to eigenstates and energies.

MoleculeSpectrum.calculate_spectrum — Function

calculate_spectrum(hamiltonian_parts, external_fields)

Compute the energies and eigenstates under the external fields.

To avoid reconstructing the Hamiltonian each time, hamiltonian_parts can be reused over calls to calculate_spectrum. The output Spectrum object is used as an input for further analysis, for example in find_transition_strengths.

See also make_hamiltonian_parts, make_krb_hamiltonian_parts, Spectrum.

MoleculeSpectrum.get_energy — Function

get_energy(spectrum, target)

Return the energy of the eigenstate with the most wavefunction overlap with target.

See also calculate_spectrum, max_overlap_with.

MoleculeSpectrum.get_energy_difference — Function

get_energy_difference(spectrum, g, e)

Return the difference in energy between the eigenstates from spectrum that have the most wavefunction overlap with g and e.

This method is used to calculate transition frequencies; g represents the ground state, and e the excited state.

See also calculate_spectrum, get_energy.

MoleculeSpectrum.find_transition_strengths — Function

find_transition_strengths(spectrum::Spectrum, g::State, frequency_range; polarization::Union{Int, SphericalUnitVector, Nothing}=nothing)

Compute electric dipole transitions out of g with energy between frequency_range[1] and frequency_range[2].

The output is a Vector of tuples (frequency, strength, closest_basis_state, eigenstate_index), produced in decreasing order of transition strength. The strength is the absolute value of the dipole matrix element, normalized by $D/\sqrt{3}$ (the maximum transition dipole between $N = 0$ and $N = 1$). We use the absolute value of the matrix element, rather than its square, so the results are proportional to Rabi frequency $Ω$.

The polarization keyword argument can be used to choose a specific microwave polarization. This defaults to nothing, which returns the incoherent sum over $σ-$, $π$, and $σ+$. If polarization is an Int, then it is interpreted as the spherical component p = -1:1 of the dipole operator (p == -1 corresponds to $σ-$ polarization). If polarization is a SphericalUnitVector, then the polarization is interpreted as linear along that axis.

There is a convenience method plot_transition_strengths that immediately produces a plot from the result.

See also plot_transition_strengths, make_hamiltonian_parts, make_krb_hamiltonian_parts, Spectrum.

MoleculeSpectrum.plot_transition_strengths — Function

plot_transition_strengths(spectrum::Spectrum, g::State, frequency_range; polarization::Union{Int, SphericalUnitVector, Nothing}=nothing)

Plot the frequencies and strengths of electric dipole transitions out of g, with energy between frequency_range[1] and frequency_range[2].

The polarization keyword argument can be used to choose a specific microwave polarization. This defaults to nothing, which returns the incoherent sum over $σ-$, $π$, and $σ+$. If polarization is an Int, then it is interpreted as the spherical component p = -1:1 of the dipole operator (p == -1 corresponds to $σ-$ polarization). If polarization is a SphericalUnitVector, then the polarization is interpreted as linear along that axis.

This method calls find_transition_strengths internally.

See also find_transition_strengths, make_hamiltonian_parts, make_krb_hamiltonian_parts, Spectrum.

MoleculeSpectrum.calculate_dipolar_interaction — Function

calculate_dipolar_interaction(spectrum::Spectrum, g::State, e::State; p::Int = 0)
calculate_dipolar_interaction(spectrum::Spectrum, g::Vector{ComplexF64}, e::Vector{ComplexF64}; p::Int = 0)

#TODO: Document calculate_dipolar_interaction

Index

MoleculeSpectrum.DEFAULT_FIELDS

MoleculeSpectrum.DEFAULT_MOLECULAR_PARAMETERS

MoleculeSpectrum.KRb_Nuclear_Neyenhuis

MoleculeSpectrum.KRb_Nuclear_Ospelkaus

MoleculeSpectrum.KRb_Parameters_Neyenhuis

MoleculeSpectrum.KRb_Parameters_Ospelkaus

MoleculeSpectrum.KRb_Polarizability

MoleculeSpectrum.KRb_Zeeman

MoleculeSpectrum.TEST_FIELDS

MoleculeSpectrum.TOY_MOLECULE_PARAMETERS

MoleculeSpectrum.ExternalFields

MoleculeSpectrum.HamiltonianParts

MoleculeSpectrum.MolecularParameters

MoleculeSpectrum.NuclearParameters

MoleculeSpectrum.Polarizability

MoleculeSpectrum.Spectrum

MoleculeSpectrum.SphericalUnitVector

MoleculeSpectrum.SphericalVector

MoleculeSpectrum.State

MoleculeSpectrum.ZeemanParameters

MoleculeSpectrum.KRbState

MoleculeSpectrum.T⁽²⁾

MoleculeSpectrum.T⁽¹⁾

MoleculeSpectrum.UnitVectorX

MoleculeSpectrum.UnitVectorY

MoleculeSpectrum.UnitVectorZ

MoleculeSpectrum.VectorX

MoleculeSpectrum.VectorY

MoleculeSpectrum.VectorZ

MoleculeSpectrum.calculate_dipolar_interaction

MoleculeSpectrum.calculate_spectrum

MoleculeSpectrum.decompose_to_basis_states

MoleculeSpectrum.find_closest_basis_state

MoleculeSpectrum.find_transition_strengths

MoleculeSpectrum.get_energy

MoleculeSpectrum.get_energy_difference

MoleculeSpectrum.get_tensor_component

MoleculeSpectrum.hamiltonian

MoleculeSpectrum.index_to_state

MoleculeSpectrum.make_hamiltonian_parts

MoleculeSpectrum.make_krb_hamiltonian_parts

MoleculeSpectrum.max_overlap_with

MoleculeSpectrum.order_by_overlap_with

MoleculeSpectrum.plot_transition_strengths

MoleculeSpectrum.state_to_index

MoleculeSpectrum.tensor_dot